+++ tests/Test/MemPackSpec.hs
@@ -19,6 +19,7 @@ import Data.Array.Byte (ByteArray)
 import Data.Bits
 import Data.ByteString (ByteString)
 import Data.ByteString.Short (ShortByteString)
+import qualified Data.ByteString.Lazy as BSL
 import Data.Complex
 import Data.Either (isLeft)
 import Data.Function (fix)
@@ -27,6 +28,7 @@ import Data.MemPack
 import Data.MemPack.Buffer
 import Data.MemPack.Error
 import Data.Ratio
+import Data.Text (Text)
 import Data.Word
 import Foreign.Ptr (IntPtr (..), Ptr, intPtrToPtr)
 import Foreign.StablePtr (StablePtr, castPtrToStablePtr, castStablePtrToPtr)
@@ -216,8 +218,8 @@ spec = do
   memPackSpec @[E Word]
   memPackSpec @(E Int, E Word)
   memPackSpec @(E Int8, E Word8, E Char)
-  memPackSpec @(E Int16, E Word16, Float, Double)
-  memPackSpec @(E Int32, E Word32, Float, Double, Ptr Char)
+  memPackSpec @(E Int16, E Word16, ByteArray, Double)
+  memPackSpec @(E Int32, E Word32, Float, ByteString, Ptr Char)
   memPackSpec @(E Int64, E Word64, Length, VarLen Word, StablePtr Char, [VarLen Word32])
   memPackSpec @(Tag, Int, Int8, Int16, Int32, Int64, Backtrack)
   memPackSpec @(E (VarLen Word))
@@ -234,7 +236,9 @@ spec = do
   memPackSpec @(Ratio (E Int))
   memPackSpec @ByteArray
   memPackSpec @ByteString
-  memPackSpec @ShortByteString
+  memPackSpec @BSL.ByteString
+  memPackSpec @ByteArray
+  memPackSpec @Text
   memPackSpec @Backtrack
   prop "Out of bound char" $ forAll (choose (0x110000, maxBound :: Word32)) $ \w32 ->
     unpack @Char (pack w32) `shouldSatisfy` isLeft